+++ CHANGELOG.md
@@ -5,13 +5,17 @@ The **goal** of this file is explaining to the users of our project the notable 
 _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)_.
 
 
-# Unreleased
+## Unreleased
 
 ### Added
 - Generallize mandatory words and add mandatory_bio_keywords
 
-# Fixed
+### Changed 
+- Update xpath for like/unlike and comment
+
+### Fixed
 - "UnboundLocalError: local variable 'commenting_approved' referenced before assignment" error when bot tries to comment
+- Typo updating configuration object. Changed nofity into notify
 
 
 ## [0.6.9] - 2020-06-12